Experiment with Nes as it's own dll
This commit shows how we could isolate the Nes code as it's own dll and
make use of it from either native or managed driver code. This involves:
1. Exporting a C style API from nes.dll i.e Nes* Nes_Create(), void
Nes_Run(Nes*), etc.
2. Fixing up pointer types and ownership inside of Nes so that an IGfx*
can be provided by the calling code. The nes dll provides an Sdl
implementaiton of IGfx that we export.

The managed driver also shows how we could defined C# bindings for nes.dll
and make use of it in a C# style. This could potentially be it's own dll
and a managed project would just reference the C# bindings dll

I played around with exporting the actual C++ class methods. This works
when the driver is also C++ and everything is being compiled with the same
compiler, but has potential issues when trying to create something more
portable. We could end up exporting both a C style and C++ style API
though. I don't know of any way to make use of C++ exports from C#. I also
don't want to do COM interop for the managed native stuff.